+++ src/neuroconv/tools/roiextractors/__init__.py
@@ -1,13 +1,13 @@
 from .roiextractors import (
     _check_if_imaging_fits_into_memory,
-    get_nwb_imaging_metadata,
-    get_nwb_segmentation_metadata,
-    add_devices_to_nwbfile,
     add_imaging_to_nwbfile,
     add_segmentation_to_nwbfile,
     write_imaging_to_nwbfile,
     write_segmentation_to_nwbfile,
 )
 from .roiextractors_pending_deprecation import (
+    add_devices_to_nwbfile,
     add_fluorescence_traces_to_nwbfile,
+    get_nwb_imaging_metadata,
+    get_nwb_segmentation_metadata,
 )
